florence base 2/requirements.txt
# Requirements pour Microsoft Florence-2
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
pillow>=10.0.0
numpy>=1.24.0
einops>=0.7.0
timm>=0.9.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.25.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Pour RunPod serverless
runpod>=1.0.0

# Note: flash_attn est optionnel et sera géré dans le code si nécessaire
